move token methods from ddsutil to oauth_utils
