+++ d4s2_auth/oauth_utils.py
@@ -1,7 +1,10 @@
 from __future__ import print_function
+from django.conf import settings
 import requests
+from ddsc.core.ddsapi import ContentType
 from requests_oauthlib import OAuth2Session
-from .models import OAuthService, OAuthToken
+from d4s2_auth.models import OAuthToken, OAuthService, DukeDSAPIToken
+from d4s2_auth.backends.dukeds import check_jwt_token, InvalidTokenError, save_dukeds_token
 import logging
 
 logger = logging.getLogger(__name__)
@@ -121,6 +124,85 @@ def revoke_token(token):
         response.raise_for_status()
     except requests.HTTPError as e:
         raise OAuthException(e)
+
+
+class NoTokenException(BaseException):
+    pass
+
+
+def get_oauth_token(user):
+    """
+    Gets the OAuth token object for the specified user, and refreshes if needed
+    :param user:
+    :return:
+    """
+    service = OAuthService.objects.first()
+    try:
+        current_user_details(service, user)
+    except OAuthException as e:
+        raise NoTokenException(e)
+    return OAuthToken.objects.get(user=user, service=service)
+
+
+def get_local_dds_token(user):
+    """
+    Gets a user's DukeDSAPIToken object if they have one
+    :param user: A django user
+    :return: The DukeDSAPIToken for the user, or None if invalid or not present
+    """
+    # If user has an existing token, check to see if it's valid
+    try:
+        for token in DukeDSAPIToken.objects.filter(user=user):
+            checked = check_jwt_token(token.key)
+            if checked:
+                return token
+    except InvalidTokenError:
+        token.delete()
+
+    return None
+
+
+def get_dds_token_from_oauth(oauth_token):
+    """
+    Presents an OAuth token to DukeDS, obtaining an api_token
+    :param oauth_token: An OAuthToken object
+    :return: The dictionary from JSON returned by the /user/api_token endpoint
+    """
+    authentication_service_id = settings.DDSCLIENT_PROPERTIES['openid_provider_id']
+    headers = {
+        'Content-Type': ContentType.json,
+    }
+    access_token = oauth_token.token_dict.get('access_token')
+
+    data = {
+        "access_token": access_token,
+        "authentication_service_id": authentication_service_id,
+    }
+    base_url = settings.DDSCLIENT_PROPERTIES['url']
+    url = base_url + "/user/api_token"
+    response = requests.get(url, headers=headers, params=data)
+    try:
+        response.raise_for_status()
+        return response.json()
+    except requests.HTTPError as e:
+        raise NoTokenException(e)
+
+
+def get_dds_token(user):
+    """
+    Returns a DukeDS api_token for the specified user, provided one can be found locally or obtained from DukeDS via OAuth.
+    Raises NoTokenException on error
+    :param user: A django user
+    :return: A DukeDSAPI token object
+    """
+    dds_token = get_local_dds_token(user)
+    if dds_token:
+        return dds_token
+    # No local token, now get from OAuth
+    oauth_token = get_oauth_token(user)
+    dds_token_json = get_dds_token_from_oauth(oauth_token)
+    dds_token = save_dukeds_token(user, dds_token_json['api_token'])
+    return dds_token
 
 
 def main():
