+++ switchboard/tests_ddsutil.py
@@ -1,11 +1,12 @@
 from django.test import TestCase
 from d4s2_api.models import DukeDSUser, DukeDSProject
 from mock import patch, Mock, MagicMock
-from switchboard.dds_util import DDSUtil, ModelPopulator, DeliveryDetails, get_local_dds_token, get_dds_token_from_oauth, NoTokenException
+from switchboard.dds_util import DDSUtil, ModelPopulator, DeliveryDetails
 from switchboard.mocks_ddsutil import MockDDSProject, MockDDSUser
 from d4s2_api.models import User
-from d4s2_auth.tests_oauth_utils import make_oauth_service
+from d4s2_auth.oauth_utils import get_local_dds_token, get_dds_token_from_oauth, NoTokenException
 from d4s2_auth.models import OAuthService, DukeDSAPIToken, OAuthToken
+
 
 class DDSUtilTestCase(TestCase):
 
@@ -88,56 +89,6 @@ class DDSUtilTestCase(TestCase):
             self.assertTrue(get_project_transfer.called_with(transfer_id))
             self.assertEqual(project_transfer.get('status'), 'accepted')
 
-
-class DDSUtilAuthTestCase(TestCase):
-
-    def setUp(self):
-        self.user = User.objects.create(username='ddsutil_user')
-        self.user_id = 'abcd-1234-efgh-8876'
-        self.token = DukeDSAPIToken.objects.create(user=self.user, key='some-token')
-        self.oauth_service = OAuthService.objects.create(name="Test Service")
-
-    @patch('switchboard.dds_util.check_jwt_token')
-    def test_reads_local_token(self, mock_check_jwt_token):
-        mock_check_jwt_token.return_value = True
-        local_token = get_local_dds_token(self.user)
-        self.assertEqual(local_token.key, 'some-token', 'Should return token when check passes')
-
-    @patch('switchboard.dds_util.check_jwt_token')
-    def test_no_local_token_when_check_fails(self, mock_check_jwt_token):
-        mock_check_jwt_token.return_value = False
-        local_token = get_local_dds_token(self.user)
-        self.assertIsNone(local_token, 'Should return none when check fails')
-
-    @patch('switchboard.dds_util.requests')
-    def test_gets_dds_token_from_oauth(self, mock_requests):
-        mocked_dds_token = {'api_token': 'abc1234'}
-        mock_response = Mock(raise_for_status=Mock(), json=Mock(return_value=mocked_dds_token))
-        mock_requests.get = Mock(return_value=mock_response)
-        oauth_token = OAuthToken.objects.create(user=self.user,
-                                                service=self.oauth_service,
-                                                token_json='{"access_token":"g2jo83lmvasijgq"}')
-        exchanged = get_dds_token_from_oauth(oauth_token)
-        # Should parse the JSON of the oauth_token and send the value of access_token to the DDS API
-        sent_get_params = mock_requests.get.call_args[1].get('params')
-        self.assertEqual(sent_get_params['access_token'], 'g2jo83lmvasijgq')
-        self.assertTrue(mock_requests.get.called)
-        self.assertTrue(mock_response.raise_for_status.called)
-        self.assertEqual(exchanged, mocked_dds_token)
-
-    @patch('switchboard.dds_util.requests')
-    def test_handles_dds_token_from_oauth_failure(self, mock_requests):
-        mock_requests.HTTPError = Exception
-        mocked_dds_token = {'api_token': 'abc1234'}
-        raise_for_status = Mock()
-        raise_for_status.side_effect = mock_requests.HTTPError()
-        mock_response = Mock(raise_for_status=raise_for_status)
-        mock_requests.get = Mock(return_value=mock_response)
-        oauth_token = OAuthToken.objects.create(user=self.user,
-                                                service=self.oauth_service,
-                                                token_json='{"access_token":"g2jo83lmvasijgq"}')
-        with self.assertRaises(NoTokenException):
-            get_dds_token_from_oauth(oauth_token)
 
 class MockDetails(object):
     full_name='Test User'
